# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CareOps AI - Inspection Readiness Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /incidents:
    get:
      operationId: listIncidents
      tags: [incidents]
      summary: List all incidents
      parameters:
        - name: childId
          in: query
          schema:
            type: integer
        - name: requiresNotification
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Incident list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Incident"
    post:
      operationId: createIncident
      tags: [incidents]
      summary: Record an incident
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncidentInput"
      responses:
        "201":
          description: Created incident with notification assessment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
  /incidents/{id}:
    get:
      operationId: getIncident
      tags: [incidents]
      summary: Get incident detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Incident detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
        "404":
          description: Not found

    patch:
      operationId: updateIncident
      tags: [incidents]
      summary: Update an incident record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncidentUpdate"
      responses:
        "200":
          description: Updated incident
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"

  # ── REGULATION 44 / 45 ─────────────────────────────────────────────────────
  /regulation44:
    get:
      operationId: listRegulation44
      tags: [regulation]
      summary: List Regulation 44 visits and actions
      responses:
        "200":
          description: Regulation 44 visits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Regulation44Visit"
    post:
      operationId: createRegulation44Visit
      tags: [regulation]
      summary: Record a Regulation 44 visit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Regulation44Input"
      responses:
        "201":
          description: Created Regulation 44 visit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Regulation44Visit"
  /regulation44/{id}:
    patch:
      operationId: updateRegulation44Visit
      tags: [regulation]
      summary: Update Regulation 44 visit or actions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Regulation44Update"
      responses:
        "200":
          description: Updated visit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Regulation44Visit"
  /regulation45:
    get:
      operationId: listRegulation45
      tags: [regulation]
      summary: List Regulation 45 reviews
      responses:
        "200":
          description: Regulation 45 reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Regulation45Review"
    post:
      operationId: createRegulation45Review
      tags: [regulation]
      summary: Record a Regulation 45 review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Regulation45Input"
      responses:
        "201":
          description: Created Regulation 45 review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Regulation45Review"
...
  /governance/complaints:
    get:
      operationId: listComplaints
      tags: [governance]
      summary: List complaints and compliments
      responses:
        "200":
          description: Complaints and compliments list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Complaint"
    post:
      operationId: createComplaint
      tags: [governance]
      summary: Record a complaint or compliment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplaintInput"
      responses:
        "201":
          description: Created complaint/compliment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"
  /governance/complaints/{id}:
    patch:
      operationId: updateComplaint
      tags: [governance]
      summary: Update a complaint record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplaintUpdate"
      responses:
        "200":
          description: Updated complaint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"
  /governance/notifications:
    get:
      operationId: listNotifications
      tags: [governance]
      summary: List Ofsted notifications
      responses:
        "200":
          description: Notifications list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OfstedNotification"
    post:
      operationId: createNotification
      tags: [governance]
      summary: Record an Ofsted notification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationInput"
      responses:
        "201":
          description: Created notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OfstedNotification"
  /governance/policies:
    get:
      operationId: listPolicies
      tags: [governance]
      summary: List policies and key documents
      responses:
        "200":
          description: Policies list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Policy"
    post:
      operationId: createPolicy
      tags: [governance]
      summary: Add a policy record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyInput"
      responses:
        "201":
          description: Created policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
...
components:
  schemas:
...
    Incident:
      type: object
      required: [id, title, date, type, severity, childId, recordedBy, notificationRequired, status]
      properties:
        id:
          type: integer
        title:
          type: string
        date:
          type: string
        type:
          type: string
          enum: [physical_altercation, property_damage, absconding, self_harm, substance_misuse, mental_health_crisis, other]
        severity:
          type: string
          enum: [minor, moderate, serious, critical]
        childId:
          type: ["integer", "null"]
        childName:
          type: ["string", "null"]
        description:
          type: string
        recordedBy:
          type: string
        notificationRequired:
          type: boolean
        notificationReason:
          type: ["string", "null"]
        notificationSuggestedWording:
          type: ["string", "null"]
        notificationSent:
          type: boolean
        riskAssessmentUpdated:
          type: boolean
        status:
          type: string
          enum: [open, under_review, closed]
        followUpActions:
          type: ["string", "null"]
    IncidentInput:
      type: object
      required: [title, date, type, severity, recordedBy, description]
      properties:
        title:
          type: string
        date:
          type: string
        type:
          type: string
          enum: [physical_altercation, property_damage, absconding, self_harm, substance_misuse, mental_health_crisis, other]
        severity:
          type: string
          enum: [minor, moderate, serious, critical]
        childId:
          type: integer
        description:
          type: string
        recordedBy:
          type: string
        followUpActions:
          type: string
    IncidentUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [open, under_review, closed]
        notificationSent:
          type: boolean
        riskAssessmentUpdated:
          type: boolean
        followUpActions:
          type: string

    # ── REGULATION ───────────────────────────────────────────────────────────
    Regulation44Visit:
      type: object
      required: [id, visitDate, conductedBy, summary, actions, overallFinding, nextDueDate]
      properties:
        id:
          type: integer
        visitDate:
          type: string
        conductedBy:
          type: string
        summary:
          type: string
        overallFinding:
          type: string
          enum: [satisfactory, requires_improvement, inadequate]
        actions:
          type: array
          items:
            $ref: "#/components/schemas/RegulationAction"
        nextDueDate:
          type: string
        reportUploaded:
          type: boolean
    Regulation44Input:
      type: object
      required: [visitDate, conductedBy, summary, overallFinding, nextDueDate]
      properties:
        visitDate:
          type: string
        conductedBy:
          type: string
        summary:
          type: string
        overallFinding:
          type: string
          enum: [satisfactory, requires_improvement, inadequate]
        nextDueDate:
          type: string
        reportUploaded:
          type: boolean
    Regulation44Update:
      type: object
      properties:
        summary:
          type: string
        overallFinding:
          type: string
          enum: [satisfactory, requires_improvement, inadequate]
        reportUploaded:
          type: boolean
    Regulation45Review:
      type: object
      required: [id, reviewDate, completedBy, period, summary, qualityRating]
      properties:
        id:
          type: integer
        reviewDate:
          type: string
        completedBy:
          type: string
        period:
          type: string
        summary:
          type: string
        qualityRating:
          type: string
          enum: [outstanding, good, requires_improvement, inadequate]
        actionsRaised:
          type: integer
        reportUploaded:
          type: boolean
    Regulation45Input:
      type: object
      required: [reviewDate, completedBy, period, summary, qualityRating]
      properties:
        reviewDate:
          type: string
        completedBy:
          type: string
        period:
          type: string
        summary:
          type: string
        qualityRating:
          type: string
          enum: [outstanding, good, requires_improvement, inadequate]
        actionsRaised:
          type: integer
        reportUploaded:
          type: boolean
    RegulationAction:
      type: object
      required: [id, title, dueDate, status, assignedTo, source]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        dueDate:
          type: string
        status:
          type: string
          enum: [outstanding, in_progress, completed, escalated]
        assignedTo:
          type: string
        source:
          type: string
          enum: [reg44, reg45, internal_audit, ofsted]
        evidenceUploaded:
          type: boolean
        completedAt:
          type: ["string", "null"]

    # ── GOVERNANCE ───────────────────────────────────────────────────────────
    Complaint:
      type: object
      required: [id, type, date, description, raisedBy, status]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [complaint, compliment, concern]
        date:
          type: string
        description:
          type: string
        raisedBy:
          type: string
        assignedTo:
          type: ["string", "null"]
        status:
          type: string
          enum: [open, under_investigation, resolved, escalated]
        resolution:
          type: ["string", "null"]
        resolvedAt:
          type: ["string", "null"]
        linkedChildId:
          type: ["integer", "null"]
    ComplaintInput:
      type: object
      required: [type, date, description, raisedBy]
      properties:
        type:
          type: string
          enum: [complaint, compliment, concern]
        date:
          type: string
        description:
          type: string
        raisedBy:
          type: string
        linkedChildId:
          type: integer
    ComplaintUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [open, under_investigation, resolved, escalated]
        assignedTo:
          type: string
        resolution:
          type: string
    OfstedNotification:
      type: object
      required: [id, type, date, description, notificationRequired, sent]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [serious_incident, missing_episode, safeguarding_concern, placement_change, manager_change, other]
        date:
          type: string
        description:
          type: string
        notificationRequired:
          type: boolean
        notificationReason:
          type: ["string", "null"]
        suggestedWording:
          type: ["string", "null"]
        sent:
          type: boolean
        sentAt:
          type: ["string", "null"]
        reference:
          type: ["string", "null"]
    NotificationInput:
      type: object
      required: [type, date, description, notificationRequired]
      properties:
        type:
          type: string
          enum: [serious_incident, missing_episode, safeguarding_concern, placement_change, manager_change, other]
        date:
          type: string
        description:
          type: string
        notificationRequired:
          type: boolean
        notificationReason:
          type: string
        suggestedWording:
          type: string
    Policy:
      type: object
      required: [id, title, category, version, lastReviewedDate, nextReviewDate, status]
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
          enum: [safeguarding, health_safety, behaviour_management, statement_of_purpose, staffing, medication, other]
        version:
          type: string
        lastReviewedDate:
          type: string
        nextReviewDate:
          type: string
        status:
          type: string
          enum: [current, due_for_review, overdue]
        owner:
          type: ["string", "null"]
    PolicyInput:
      type: object
      required: [title, category, version, lastReviewedDate, nextReviewDate]
      properties:
        title:
          type: string
        category:
          type: string
          enum: [safeguarding, health_safety, behaviour_management, statement_of_purpose, staffing, medication, other]
        version:
          type: string
        lastReviewedDate:
          type: string
        nextReviewDate:
          type: string
        owner:
          type: string
...